src/requirements.txt
psycopg2==2.7.6.1
django==2.1.3
djangorestframework==3.9.0
django_filter==2.0.0
django_extensions==2.1.3
django-s3-storage==0.12.4
geojson==2.4.1
drf-yasg==1.11.0
zappa==0.47.0
